refactor: 모달에 관련된 컴포넌트를 분리
